Rename path utils module
