+++ src/archivematica/storage_service/vue/lib/location-directory-picker/index.ts
@@ -1,7 +1,7 @@
 import { createApp, defineComponent, h, ref } from 'vue'
 import App from './App.vue'
 import { i18n, initI18n } from '@/shared/i18n'
-import { toAbsolutePath, toRelativePath } from './pathUtils'
+import { toAbsolutePath, toRelativePath } from './utils/path'
 import 'font-awesome/css/font-awesome.min.css'
 
 async function bootstrap() {
