+++ src/archivematica/storage_service/vue/lib/location-directory-picker/utils/path.ts
